gralloc: make gralloc_drm_handle_t layout consistent on 32/64-bit
gralloc_drm_handle_t's memory layout must be consistent between the 32-bit
and 64-bit gralloc, in case buffers are passed between processes.
Replace the (variably sized) int data with a fixed size uint64_t,
and enforce 8-byte alignment for architectures where uint64_t alignment
varies between 32-bit and 64-bit

+++ gralloc_drm.c
@@ -211,7 +211,7 @@ static struct gralloc_drm_bo_t *validate_handle(buffer_handle_t _handle,
 		}
 
 		handle->data_owner = gralloc_drm_get_pid();
-		handle->data = (int) bo;
+		handle->data = (intptr_t) bo;
 	}
 
 	return (struct gralloc_drm_bo_t *) handle->data;
@@ -296,7 +296,7 @@ struct gralloc_drm_bo_t *gralloc_drm_bo_create(struct gralloc_drm_t *drm,
 	bo->refcount = 1;
 
 	handle->data_owner = gralloc_drm_get_pid();
-	handle->data = (int) bo;
+	handle->data = (intptr_t)bo;
 
 	return bo;
 }
